fix(internal): give the companies list and board the same header
- Each view built its own header and they had drifted apart — the board's title was even rendering in the plain body font because its header styles weren't loading. One shared header now fixes the styling and keeps both views consistent.
- Moved the list/board switch into that shared header so it sits in the same place on both views, rather than tucked inside the list's filter bar.
